+++ protocol/clproto_cpp/test/tests/test_messages.cpp
@@ -22,8 +22,6 @@ TEST(MessageProtoTest, EncodeDecodeState) {
   EXPECT_EQ(send_state.is_empty(), recv_state.is_empty());
   EXPECT_EQ(send_state.get_type(), recv_state.get_type());
   EXPECT_STREQ(send_state.get_name().c_str(), recv_state.get_name().c_str());
-  EXPECT_EQ(send_state.get_timestamp().time_since_epoch().count(),
-            recv_state.get_timestamp().time_since_epoch().count());
 }
 
 TEST(MessageProtoTest, EncodeDecodeInvalidState) {