+++ python/test/test_clproto.py
@@ -49,7 +49,6 @@ class TestClprotoJSON(unittest.TestCase):
     def test_string_comparison(self):
         msg = clproto.encode(sr.CartesianPose("A", 1.0, 0.5, 3.0, "B"), clproto.MessageType.CARTESIAN_POSE_MESSAGE)
         json = clproto.to_json(msg)
-        json = json[:json.find(",\"timestamp")] + json[json.find("},\"referenceFrame"):]
         self.assertEqual(json, "{\"cartesianPose\":{\"spatialState\":{\"state\":{\"name\":\"A\",\"type\":"
                                "\"CARTESIAN_POSE\"},\"referenceFrame\":\"B\"},\"position\":{\"x\":1,\"y\":0.5,"
                                "\"z\":3},\"orientation\":{\"w\":1,\"vec\":{}}}}")
@@ -58,14 +57,12 @@ class TestClprotoJSON(unittest.TestCase):
         joint_state.set_velocities([0.3, 0.1, 0.6])
         msg = clproto.encode(joint_state, clproto.MessageType.JOINT_STATE_MESSAGE)
         json = clproto.to_json(msg)
-        json = json[:json.find(",\"timestamp")] + json[json.find("},\"jointNames"):]
         self.assertEqual(json, "{\"jointState\":{\"state\":{\"name\":\"robot\",\"type\":\"JOINT_STATE\"},"
                                "\"jointNames\":[\"joint0\",\"joint1\",\"joint2\"],\"positions\":[0,0,0],"
                                "\"velocities\":[0.3,0.1,0.6],\"accelerations\":[0,0,0],\"torques\":[0,0,0]}}")
 
         msg = clproto.encode(sr.Jacobian("robot", 3, "test"), clproto.MessageType.JACOBIAN_MESSAGE)
         json = clproto.to_json(msg)
-        json = json[:json.find(",\"timestamp")] + json[-3:]
         self.assertEqual(json, "{\"jacobian\":{\"state\":{\"name\":\"robot\",\"type\":\"JACOBIAN\",\"empty\":true}}}")