+++ source/state_representation/include/state_representation/State.hpp
@@ -90,15 +90,14 @@ public:
   void set_filled();
 
   /**
+   * @brief Get the age of the state, i.e. the time since the last modification
+   */
+  double get_age() const;
+
+  /**
    * @brief Getter of the timestamp attribute
    */
   const std::chrono::time_point<std::chrono::steady_clock>& get_timestamp() const;
-
-  /**
-   * @brief Setter of the timestamp attribute
-   * @param timepoint the new value for the timestamp
-   */
-  void set_timestamp(const std::chrono::time_point<std::chrono::steady_clock>& timepoint);
 
   /**
    * @brief Reset the timestamp attribute to now
@@ -114,6 +113,12 @@ public:
    * @brief Setter of the name
    */
   virtual void set_name(const std::string& name);
+
+  /**
+   * @brief Check if the state is deprecated given a certain time delay
+   * @param time_delay the timer after which to consider the state as deprecated
+   */
+  bool is_deprecated(double time_delay);
 
   /**
    * @brief Check if the state is deprecated given a certain time delay